Estimation of the Half Maximal Inhibitory Concentration

The half maximal inhibitory concentration (IC50) is a measure of the effectiveness of a compound in inhibiting biological or biochemical function. This quantitative measure indicates how much of a particular drug or other substance (inhibitor) is needed to inhibit a given biological process (or component of a process) by half.

See reference(s).

In this version of the function the maximal occupancy (rmax) is estimated automatically. This should be optional.

hillEquation(conc, occ, guess = c(1, 100),
  control = minpack.lm::nls.lm.control())

Arguments

conc
a vector of drug concentrations in plasma (example units are ng/mL).

occ
a vector of PET occupancy values that correspond to the measured drug concentrations in plasma.

guess
a length-two vector of starting values for the nonlinear optimization.

control
is a list of parameters used by nls.lm.control that are set by default, but may be customized by the user.

Value

List with the following elements

IC50Half maximal inhibitory concentration

rmaxEstimated maximal occupancy

IC50SEApproximate standard error for IC50

rmaxSEApproximate standard erorr for rmax

hessianHessian matrix from the Levenburg-Marquardt procedure

infoReturn value from the Levenburg-Marquardt procedure

devianceDeviance from the Levenburg-Marquardt procedure

messageText message from the Levenburg-Marquardt procedure

Author

Brandon Whitcher bwhitcher@gmail.com

References

Hill Equation IC50

See also