The Simplified Reference Tissue Model

The simplified reference tissue model (SRTM) estimates the binding potential from an observed time activity curve without the need for aterial sampling. It assumes a one-tissue compartment model to describe the influx and efflux in the tissue region of interest and the reference region.

simplifiedReferenceTissueModel(tac, ref, time, SRTM2 = TRUE, k2prime = NULL,
  guess = c(R1 = 0.5, k2 = 0.01), control = minpack.lm::nls.lm.control())

Arguments

tac
a vector corresponding to the time activity curve from the tissue (in Bq/mL).

ref
a vector corresponding to the time activity curve from the reference region (in Bq/mL).

time
a vector of average frame times (in minutes).

SRTM2
a logical value that selects the three-parameter model (SRTM) or the two-parameter model (SRTM2), where k2prime is fixed.

k2prime
the value of k2prime that has been fixed.

guess
values for the inital parameter estimates for R1 and k2.

control
a list of parameters used by nls.lm.control that are set by default, but may be customized by the user.

Value

BP

Binding potential

R1

Ratio of the volumes of distrubution for the tissue and reference region

k2

Clearance rate constant from the tissue to plasma

BP.error

Approximate standard error of the binding potential

R1.error

Approximate standard error for the ratio

k2.error

Approximate standard error for k2

Details

See the references.

The model has been parameterized in the manner of Wu and Carson (2002). That is, the nonlinear regression estimates R1, k2 and k'2 for the three-parameter model (SRTM) and R1 and k2 for the two-parameter model (SRTM2).

The convolution is performed after interpolating the time activity curves, both for the tissue and the reference region, to one-second resolution then downsampling them back to the original sampling rate.

Author

Brandon Whitcher b.whitcher@gmail.com

References

Lammertsma, A.A. and Hume, S.P. (1996) Simplified reference tissue model for PET receptor studies, NeuroImage, 4, 153-158.

Wu, Y. and Carson, R.E. (2002) Noise reduction in the simplified reference tissue model for neuroreceptor functional imaging, Journal of Cerebral Blood Flow \& Metabolism, 22, 1440-1452.

See also