Compute Drug Occupancy with Approximate Standard Errors

Receptor occupancy is calculated from posititron emission tomography (PET) data as the treatment-induced relative change in the concentration of available (not occupied) receptors.

occupancy(base, drug, baseSE = NULL, drugSE = NULL, base.drug.corr = 0)

Arguments

base

is the baseline binding potential (BPND).

drug

is the post-treatment binding potential (BPND).

baseSE

is the standard error for the baseline BPND.

drugSE

is the standard error for the post-treatment BPND.

base.drug.corr

is the user-specified correlation between baseline and post-treatment binding potentials.

Value

OCC

is the percent drug occupancy.

SE

is the approximate standard error of the parameter estimate.

Details

Occupancy is calculated using the straightforward and well-known formula. If the standard errors for the two binding potentials are provided, then the delta method is used to approximate the standard error for the estimate of occupancy.

References

Cunningham VJ, Rabiner EA, Slifstein M, Laruelle M (2010). Measuring drug occupancy in the absence of a reference region: the Lassen plot re-visited, Journal of Cerebral Blood Flow & Metababolism, 30, 46-50.

Passchier J, Gee A, Willemsen A, Vaalburg W, van Waarde A (2002). Measuring drug-related receptor occupancy with positron emission tomography, Methods, 27, 278-286.

See also

Author

Brandon Whitcher bwhitcher@gmail.com