Plot Baseline Versus Post-Treatment Binding Potentials

Inspired by the Lassen plot (Cunningham et al., 2010) this is a straightforward graphical summary of pre-treatment versus post-treatment binding potentials for a single subject across multiple brain regions.

plotBindingPotential(
  base,
  drug,
  lty45 = 2,
  lty = 1,
  lwd45 = 2,
  lwd = 3,
  col45 = "darkgrey",
  col = "orange",
  pch = 1,
  cex = 1,
  xlim = range(0, base, 0.5),
  ylim = range(0, drug, 0.5),
  xlab = expression(BP[ND]^{
     Base
 }),
  ylab = expression(BP[ND]^{
     Drug
 }),
  ...
)

Arguments

base

is the vector of baseline binding potentials across brain regions.

drug

is the vector of post-treatment binding potentials across brain regions.

lty45

is the line type for the 45-degree line.

lty

is the line type for the estimated regression line.

lwd45

is the line width for the 45-degree line.

lwd

is the line width for the estimated regression line.

col45

is the color for the 45-degree line.

col

is the color for the estimated regression line.

pch

is the plotting character symbol.

cex

is the size of the plotting symbol.

xlim

is the range of values on the x-axis.

ylim

is the range of values on the y-axis.

xlab

is the label on the x-axis.

ylab

is the label on the y-axis.

...

additional arguments to be passed to the plot function.

Value

A plot is shown, NULL is returned

Details

See the reference below.

References

Cunningham VJ, Rabiner EA, Slifstein M, Laruelle M (2010). Measuring drug occupancy in the absence of a reference region: the Lassen plot re-visited, Journal of Cerebral Blood Flow & Metababolism, 30, 46-50.

See also

Author

Brandon Whitcher bwhitcher@gmail.com